2.6.4-beta2
Added drag and drop in Bars and Icon Rows that will allow for arranging to swap around.
Fixed State field in the Info Area to respect Binary Show As
Improved Bars logic and hopefully will remove - bars in favor of - bar_x format for better yaml and uniqueness. Added migration to fix bar issues of not showing.